+++ powershell-awesome-framework/snippets/user/PAF_Get-ProcessInfo.ps1
@@ -1,15 +1,15 @@
 <#
-.Synopsis
+.SYNOPSIS
 This snippet get processes info.
 
-.Description
+.DESCRIPTION
 This script demonstrates how to retrieve basic system information using WMI and CIM cmdlets.
-
-.Category
+#>
+function Get-ProcessInfo {
+    <#
+:CATEGORY
 System Information
 #>
-
-function Get-ProcessInfo {
     [CmdletBinding()]
     param (
         [Parameter(Mandatory = $false, Position = 0)]
@@ -19,16 +19,16 @@ function Get-ProcessInfo {
     try {
         $process = Get-Process -Name $ProcessName -ErrorAction Stop
         $processInfo = @{
-            Name = $process.ProcessName
-            ID = $process.Id
-            Path = $process.Path
-            Company = $process.Company
-            CPU = $process.CPU
-            Memory = $process.WorkingSet64 / 1MB
-            Description = $process.Description
-            StartTime = $process.StartTime
+            Name               = $process.ProcessName
+            ID                 = $process.Id
+            Path               = $process.Path
+            Company            = $process.Company
+            CPU                = $process.CPU
+            Memory             = $process.WorkingSet64 / 1MB
+            Description        = $process.Description
+            StartTime          = $process.StartTime
             TotalProcessorTime = $process.TotalProcessorTime
-            Responding = $process.Responding
+            Responding         = $process.Responding
         }
 
         return $processInfo